+++ guillotina/db/storages/dummy.py
@@ -1,4 +1,4 @@
-from guillotina.db import ROOT_ID
+from guillotina.const import ROOT_ID
 from guillotina.db.interfaces import IStorage
 from guillotina.db.storages.base import BaseStorage
 from guillotina.exceptions import ConflictIdOnContainer